Add script to find all part numbers in a schematic by part triple.
This script matches partial triples, so you can search for "0402-10K",
and get all of the resistors, horizontal or not, that are 0402/10K, etc.

I wrote this to make comparing between schematics easier so that we can
optimize our component choice across all designs since we'd rather have
fewer unique parts to order, to take advantage of volume pricing.